solenoid: optimised and fixed solenoid controls
The solenoid user controls had not been updated properly after the major
changes to the autonomous system.

The main issue that was fixed was that the tick delay had not been
updated according to the new tick-speed, so the delay had become 5x
shorter.

An optimisation was also possible through only updating the solenoid
state only once a mutation had occured to it.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>