auton: further routine improvement and quick-fixes
The autonomous rountines have been changed due to them messing up the
odometry system when the robot moves forwards for 1.2 seconds, as that
is far too long and causes the odometry wheel sensors to do more than
full rotation, which, in the current implementation of things, cancels
out the previous movement; rendering the odometry system inaccurate.

I've solved this issue by shrinking the gap in the time by an order of
magnitude, though this will need to be tuned as if it's too short, it
won't properly grab hold of the mogo, and if it's too long then it will
mess with the odometry system.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>