auton routine: added an quick experimental solution
I've implemented an experimental solution to the issue with odometry when
grabing the mogo, where, instead of making the movement small enough to
not cause a full rotation, the robot will simply move backwards for the
same duration, and hopefully, this will counter-act the forwards movement
within the accuracy of a revolution of the odometry wheels; from which
the odom can kick back in to correct any smaller errors.

This is untested, though logically it *should* check out.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>