auton routine: improved the routine during nationals
Through real-life testing at nationals, I've improved and also
optimised the autonomous routine.

The autonomous routine was also changed as i had not previously
realised that the donuts displayed in the picture of the field
were actually of the opposing team's colour. So instead, the robot
now reaches the mogo, turns and grabs the bottom ring of the stack
to score.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>