auton routine: hopefully fixed the routine for finals
The issue with the last game of today was that the robot had not
stopped at where I had wanted it to stop and in turn pinned the
mogo against the pylon at high speeds which flung it in the wrong
direction.

It turned out that, it was due to the autonomous instruction itself
being wrong, rather than any issues with the new PID code. The autonomous instruction had asked for a forwards movement of 120cm,
which, after moving that distance, left no physical space (2cm) for
the mogo.

I've solved this issue through simply not moving 120cm and instead stopping far short of that and grabbing the second ring at an angle.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>