2025 BotGoBrrr v7: a fresh start
This year I've decided to write the robot code in a mix of zig & arm asm
instead of rust. Due to that, I've cleared all the files in this repo in
preparation for a new start/rewrite/major version. The reason why I
haven't just made a new repo instead is merely for historical git
history.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>