added odometry, a port buffer & better port safety
Added an __errno stub to let tests compile.
The zig.mk file was also modified so that it produces release binaries.

PROS error handling has been completely replaced by a port_buffer system
in which, there is are bitflags / a bitfield of all the ports on the
robot brain (which takes up 3 bytes of memory) that stores whether or
not the ports are working or not (rather if any errors were encountered)
and each tick this buffer is then written to a file so that later, if
issues arise, I can figure out if it's a hardware issue or a bug.

I've also added some better port safey-checking and also pros error
handling (as mentioned above).

Basic odometry has also been implemented through a global coordinate
system that's implemented through vectors and vector maths in which
every tick, the global coordinate is added to by the estimated movement
vector that's approximated through using the current calculated change
in the robot's magnitude (gotten through averaging the two rotational
sensors on each robot diagonal), at the current angle the robot is at.

Tests have also been written for odom.

Also I've switched to radians as they're easier to work with.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>