port buffer: external analysis tool & stdout logs
I've added some quick optional logging to stdout for port disconnects
through the port buffer that simply checks if there are any ports that
are disconneted and then prints them to stdout, and only when the robot
is run without a competition switch.

I've also written a simple port buffer binary file analysis tool that
just collects ALL the ports that've disconneted in the life-time of the
program runtime (through a simple bitwise and) and also the final state
of the robot (in case something disconnects and then reconnects). This
could be extended by including a calculated timestamp alongside the
disconnect.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>