added simple position recording to stdout & file
I've added a quick and simple way to record taret robot coordinates /
paths by simply having the robot log it's current position to both
stdout & a specified text file whenever a specified button is hit (x for
now).

Also some bugfixes were written as the previous commit had not been
tested to see if it even compiled or not.

Odom was moved to make it look nicer.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>