added auton framework, vector math & optimized odom
I've written a basic autonomous 'framework' if you will, for how the
auton code *should* execute. I've chosen to use a sort of state-machine
with an array of path 'points' that the robot should go to. This was
chosen instead of having a non-looping routine that had many little mini
loops within it. I've chosen this, as it's much easier to keep timings
and all the functionality 2d and simple while remaining 'concurrent'.

First the robot will update odometry, then if it hasn't already reached
it's goal, then it'll try reach it for this tick.
If it's already reached it, then it would execute any relavent tasks
(using a state variable for any multi-tick tasks) for this tick.
And then it'll update the next goal based on the planned path.

Here's an example of possible multi-tick task executions using
a state-machine:
```zig
  var task_ticks_remaining: u8 = 12;
  while (true) {
    if (condition) task_ticks_remaining = 4;

    while (task_ticks_remaining > 0) {
      // execute some task for a single tick
      task_ticks_remaining -= 1;
    }
  }
```

I've also moved a bunch of functions from odom into their own more
generic "vector" file, with more linear algebra functions that work
on any vector, not just coordinates as per used in odometry.

I've also changed all the f32 values to f64, as the precision is
REQUIRED for the calculations I anticipate (perhaps MPC or pure pursuit).

I've also optimized the odom 'minimalAngleDiff' to use a simple if
statement instead of the prettier but extremely slow sin counter-part.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>