added motor configs and removed redundant code
Changes made:
  * (-) redundant old motor port creation functions
  * (+) proper motor config struct
  * (+) proper drivetrain motor
  * (+) init code
  * (-) redundant code zig already has
  * (~) misc bugfixes

This code was initally written well over 3-4 months ago, when I was
incredibly new to zig.

So anyhow, I've reviewed most of the code that's been written, and
removed any code that just seems redundant in that zig already has a far
more optimized version of it, or there's just no need for it, as
logically they are issues that would realistically present.
I've also fixed some random bugs that I came across (without testing).

I've also added proper motor configs for the drivetrain as I plan to use
the motor's built-in pid with the `move_velocity` and also read the
motor encoder ticks that depend on the aformentioned motor configs.

I've removed a LOT of comptime that's realistically useless.

There's also now actual intialization code (for motors only so far).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>