reduced rotation sensors used in odom to only be one sensor
Changes made:
  * (~) switched to a single rotation sensor for odom

Recently, I've been notified of a change to how the robot is to be built.
In that, instead of two rotation sensors, on the top left and bottom
right wheels, there will be one rotation sensor in the middle of the
robot (as per the meta).

This makes my life SIGNIFICANTLY easier, and simplifies the calculations
as I no longer need to average the rotation sensor readings.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>