odom logging: added odom velocity & coord logging
Changes made:
  * (+) odometry coordinate CSV logging (every 160ms)
  * (+) movement & rotational velocity measurements to odom
  * (~) bugfix to updateOdom so that the previous measurements are updated
  * (+) optional logging of the movement & rotational velocities (every tick)

I've copy and pasted the code I'd had for logging motor temperature but
for odom coordinates instead.

I've also added consistent movement and rotational velocity measurements
/calculations to odom, so that the velocities for future controllers
(such as the 'MAJOR minor™' controller I'm planning to implement).
This info is calculated with the odom code, as it's already calculating
the coordinate, so it's a neat and clean location to store it all so we
have consistent calculations.

I've also noticed a bug in the updateOdom function, in that, it doesn't
actually update the last values, so it only compares with the start of
the program instead of the last incovation, which is not what I want.
I've fixed it.

I've also added logging for the newly calculated velocities. In excel
I can make two different tables from this data.
I'm grouping them together for convenience reasons.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>