logging: fixed some logging issues
Changes made:
  * (~) made logs log to the actual correct file
  * (~) switch unicode degrees symbol for `*`
  * (~) getYaw & getRotation now return 0 instead of NaN when there are errors

I've made some bugfixes to the code, where before, as I copy-pasted it,
the other logging functions were polluting the drive temperature logfile
as they were logging to it, instead of their actual individual logfiles.

Also, CSV does not support UTF-8 symbols like the degrees symbol, so
I've replaced it with something that looks similar (*).

Also, looking through the logs of the odom velocity, and coordinates,
I've noticed a large number of NaNs, this is due to the testing setup
I had only having one motor, and no sensors.
That's when I realized, that, the getFoo functions, even though
reporting the errors to the port buffer, still return the invalid number
(NaN) as the recorded value.
I've replaced it, so that it returns 0 instead, so it looks nicer.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>