PID: added a basic PID with improvements
Changes Made:
  * (+) pid.zig module
  * (+) basic PID with integral anti-windup and a low-pass filter

I've implemented a basic PID for the robot with the standard
improvements (integral anti-windup & a low-pass filter).

I've split the parameters and state into different structures, as the
parameters are never modified, and therefore need not be part of the
'state' of the PID that gets updated.
This is also to allow for easy and quick resetting of the state of the
PID without having to re-input it's parameters.

I've seen some posts online that say that a derivative in a PID
controller is not needed when trying to set the velocity of there exists
no 2nd order inertia.
We're try to set/control the distance however, so that doesn't apply to
our main use-case.

The next course of action, is to implement ample logging of the PID
error term so tuning is a breeze.
Also include logging of the prev_derivative, so we can tune to reduce
noise for the derivative (through low_pass_a).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>