akibot modifications
I've decided to make a new branch, with modifications to the code, that
allows it to work with 'akibot'- out school's other robotics team.

This is as, I have VERY limited physical access to OUR bot, so I want
the option to test the code I've written on the other team's bot.

These are just minor modifications to make it compatible with their bot.

Also, their bot doesn't have rotation or imu sensors for odometry, so
to suffice, I've chosen to just use the motor encoders of the top left
motor of their bot, and for the imu, it's left untouched so the robot
will be recorded as always facing 'forwards' (angle 0).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>