logging: moved logging into it's own file
Changes made:
  * (+) added the logging.zig file
  * (~) moved all drive motor temp & odom logging to logging.zig
  * (~) renamed logging func names to not be redundant
  * (~) moved updateOdom into the odom state, and simplified it's name (simplicity)
  * (+) added battery temperature logging

I've chosen to move all the logging into a logging.zig instead, as some
sorts of logs, such as temperatures, would greatly benefit from logging
from multiple different sources such as drive motors, arm motors and
battery temperatures, which doesn't make sense to put into one of the
files of any individual component.

Also, I've chosen to simplify the function names to not be redundant
and moved the updateOdom function into the odom state, as it would
never operate upon anything but that state, so this is for cleaner
and simpler code. OOP is not evil.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>