minor opcontrol split arcade tweak
Changes made:
  * (~) changed drivetrain motor reverseness to fit how the robot physically works
  * (~) tweaked split arcade to actually work and feel nice

Recently, after testing split arcade, I noticed that due to how I
programmed it, the x and y axes were swapped. With the left controller
y, controlling the x of the robot, and the right controller x
controlling the y of the robot (if you were to look at the previous
implementation).

I initially just fixed this, by swapping them, but it didn't feel right
to have the left joystick be the y, so now I've made it so that the
left joystick is the x and the right joystick is the y.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>