logging simplification + benchmark & battery
Changes made:
  * (+) battery & controller battery percentage logging
  * (+) optional logging of cycle times for benchmarking
  * (+) added helper simple helper functions for handling optional files
  * (~) simplified logging code by making all logging functions handle optional files
  * (~) added time to the odom coordinate log

I've added controller and battery percentage logging, so we can track
and calculate how fast the batteries drain, so we can know when to
recharge, and how many games each battery or charge will last for.

I've also added *optional* (as goes with any logging that happens every
tick) logging of cycle times for benchmarking. I record the time it
takes to do the actual computations of the cycle and time it takes to
log separately as I'm guessing that IO is INCREDIBLY slow. For
simplicity reasons, the logging of the benchmark itself, is not included
in the logging.
Benchmarking will allow me to see, how much leeway I actually have in
terms of compute time, so I may be able to reduce the cycle delay.

I've noticed that logging currently is incredibly verbose and repetative
, with an ungodly amount of boilerplate code- most of it due to the
special handling of optional files.
So, to solve this issue, I've just added a few helper functions, and
forced all logging functions to work with optional files (even though
they would still do their job without it, even though it's not their
responsiblity). This is fine, as the logging functions will NEVER be
used without an optional file, this is an assumption we can make, and
a piece of information we can KNOW and make static.

I've also added time to the xy plot, in case you wish to compute
velocity from that somehow.
I'd ommitted it previously, as I'd wished to use it for an x y scatter
plot, and I still can, just by not selecting the time field.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>