changed motor.zig into Motor.zig + move_velocity
Changes made:
  * (~) renamed motor.zig -> Motor.zig
  * (~) changed Motor into it's own self-contained struct
  * (+) added a move_velocity function
  * (~) moved port buffer writing logic into the Motor struct

I've made a LOT of simplifications to the motor code, to make it FAR
less verbose by using an OOP design style.
Previously I've been rather adversed to OOP, due to the unnecessary
coupling of code and unmaintainability it includes.

But this is okay, as realistically, it will never need to be decoupled,
as for the most part, all the data stored in the struct, is grouped
together based upon it's usage, based upon the calculations that need
them.
And also, the information in this struct, will never be needed
outside of motor operations, and therefore, the data contained within
can be hidden by the abstraction without issue.
It's also rather static, as everything is comptime, and if you remember
to think of it, only in terms of stored configurations.

I've made the move velocity function take in a double from 0-1, since
the rpm depends on the motor's gearset and I don't wish to have to
handle that externally.

I've also moved the port buffer writing logic into the function itself,
as the boolean value it returned would have never been used for anything
other than writing to the port buffer.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>