added tower support & logs and cleaned up opcontrol
Changes made:
  * (+) tower support (tower.zig)
  * (+) tower motor temperature logging
  * (+) tower opcontrol controls
  * (~) moved drive controls out of opcontrol and into drive.zig
  * (~) made certain controller controls settable in the top-level container

The design of our robot this year, includes a tower that contains two
5.5W motors that spin in the same direction (set 200RPM).
So I've implemented rudimentary support for it, alongside logging of
it's motor temperatures.

I've set the controls to L1 and L2 as they just make sense.

I've chosen to spin the towers based upon velocity instead of voltage,
as that allows for more consistency, and the ability to spin it at a
lower speed without impeding torque.

I've cleaned up the driver control loop even further, by moving the drive
control logic, into the drive.zig file itself.
As the drive logic code, affects the drivetrain only, and the rest of the
opcontrol loop needs not know of it.

I've also made certain controller controls configurable on the top-level
container for convenience reasons, as now they're just enum values
(different from the old rust pros-rt library).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>