drive: DMN, velocity, no f32 and reverse driving
Changes made:
  * (~) switched from setting the -12000..=12000 voltage to -1..=1 velocity in the drivetrain
  * (~) purged all single-precision floats from the drive code
  * (~) revamped daniel's magic number to be within -1..=1 and calculate itself
  * (+) added an option for daniel's magic number to be applied
  * (+) added support for togglable reverse driving

I've chosen to switch from setting the voltage of the drivetrain motors,
to instead setting the velocity and using the motor's internal PIDs.
This is so, that the motor speeds are more consistent, regardless of
load, and also allows for the turning speed to be reduced without
impeding motor power/torque.

I've purged all single-precision (f32) floats from the drive code, as
f64's are far more precise with basically no performance decrease as
they are scalar values (vectors are a different story).

I've also revamped Daniel's Magic Number so that it's within the range
of -1..=1 instead of -12000..=12000 and added an option to enable it
(have it applied to the arcade drive).
Note how the multpliers must be applied AFTER daniel's magic number, as
to preserve the full curve.
What this means, is that, if you apply the multipliers after DMN, the
curve is simply shrunk down to that new scale, while if you do it before
DMN, then you only sample a small portion of DMN, which makes it closer
to being linear.

I've also added support for driving backwards, as this year's game
demands quite a lot of it.
I've chosen to make it a toggle instead of a hold-down sort of control,
as large portions of the game are going to be played whilst driving
in reverse, and it would feel awkward to have to hold down a button
the entire time.
The chosen button is X as the l1,l2 buttons are taken by the tower,
and the r1 and r2 buttons are taken by a future toggle_arcade feature.
Vibrations are added as a haptic reminder to drivers, in case they
accidentally hit it and or forget if they're going back to normal drive
or reverse drive.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>