tuner: added PID tuning options & many bugfixes
Changes made:
  * (+) added tuner.zig for optional tuning runtimes
  * (+) added the `-Dtune=` `mov-pid` & `yaw-pid` options
  * (~) changed auton runtime cycle/tick delay to 10ms (from 50ms)
  * (-) cleaned up the old deprecated auton state-machine
  * (+) added auton wait function that also updates odom
  * (+) added untuned movement & yaw PID parameters to auton
  * (~) split the arcadeDrive function into user & auton variants
  * (~) optimized odom magnitude calcuation
  * (#) removed erroneous extra degree radian conversion in odom getRotation
  * (#) added odom imu reset/init to initialize
  * (#) renamed fd to kd in PID (spelling mistake)
  * (~) added PID equation disambiguation comment
  * (#) switched to std.math.clamp instead of @min for clamping PID output to saturation
  * (+) added tuneMovPID & tuneYawPID PID tuning/logging/graphing runtimes

I've added a tuner.zig file, so to make the job of tuning the various
PIDs used in the project much easier for myself.
What I wanted, was a simple PID tuning scenario, a simple step response
and a graph of the plant output, and the desired setpoint, so that I can
easily and visually see any overshoots, undershoots and if the PID is
just not responsive enough, so that I can tune it effectively.

I've chosen to use one `tune` build option that takes in a string id
instead of multiple build options, as I expected to implement a bunch
of tuning runtimes and I do not wish to have to specify them all
manually in the build.zig file, nor do I wish to have to check for
all of them individually in the already bloated main.zig opcontrol
function.

The auton tick cycle has been changed from 50ms to 10ms, as it's
suggested to never go above nor below that time, as above would
introduce unacceptable inaccuracies, and running under that time would
waste compute, as the sensors only update every 10ms.

As now I've redesigned how auton is going to work, in that it'll be a
linear function that calls state machines when it needs to move to
certain places instead of a state machine that'll perform tasks when
certain conditions are met.
I'm switching to this, due to the sheer simplicity, maintainability and
efficientcy it provides.
It also allows for me to use a multitude of robot control stategies
instead of just one, in case I need some sort of specialist movement
(right tool for the job).

Anyhow, I've cleaned up the auton code to remove the old dynamic
state-machine system, and also added a wait function, as the first
taste of linear code calling dynamic state-machines, where the wait
function will wait for a time whilst still updating odom.

Whilst writing the tuning runtimes for the yaw and movement, I ran into
the issue that I needed PIDs to graph in the first place.
So I quickly wrote some basic untuned PIDs in autonomous.zig that used
the max_error assumption for kp and also some random correlations for
ki and kd that can be graphed and *then* tuned.

Also, I needed to use the arcade drive function for auton and tuning,
in which I discovered that I'd made some modifications to it
(DMN & turning multiplier) for user control, that I didn't want for
PID tuning nor autonomous.
So, I've split the arcadeDrive into two different functions, one for
auton and the other for opcontrol.

The rest of the changes, are just minor (but affects a LOT) bugfixes
that I caught whilst stepping through and reading all the code from init
to end (for tuning at least).
As, I have NO time whatsoever to test on the robot, and I've gotta do as
much as I can ahead of time.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>