major minor control: implemented the basic idea
Changes made:
  * (+) major_minor.zig
  * (+) untuned major minor coast rates & integrals
  * (+) major minor move function
  * (+) major minor rotate function
  * (+) added dotproducts to vector.zig
  * (~) optimized std.math.pow(x, 2) to @exp2(x)
  * (~) optimized std.math.sqrt(x) to @sqrt(x)

I've implemented a controller, I'll call the 'Major Minor Control'.
The basic idea of the controller, is with how it handles overshoots.

Each movement/rotation the robot takes, has a 'major' and 'minor'
component, with the 'major' component being the main action of the robot
, the action it'll try complete at fast as possible and if there's an
overshoot, the robot just ignores it and moves on.

The second component, is the minor control, the job of the minor control
is to correct the overshoots and errors of the previous major movement,
so as the robot is performing the next major movement, it's also
correcting for the overshoots and errors of the previous major movement.
This is done through an integral, and there are arbitrary tuning guides
to make my life simpler.
(If the major movement was proportional, then you could call it a split
P I controller)

How major movement is implemented, is rather arbitrary, but the current
idea is through the use of coasting-rates.
This makes the assumption, that friction is uniform and constant, and
that the robot will coast without rotating (each side has the same
amount of friction).
The idea is, that, from the current velocity and the rate of
deceleration when the robot coasts (no voltage is provided to the motors)
you can calculate the distance the robot will travel even after there is
no voltage applied to the motors.
And from this distance, you can subtract it from your error, and stop
just before and coast to your destination.

Eg. The robot slides/coasts for 2cm after stopping, you want to travel
    10cm, so you travel 8cm and stop and you'll reach 10cm after the
    slide/coast.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>