pid: added pid state machine movement functions
Changes made:
  * (+) PID 'move' state-machine
  * (+) PID 'rotate' state-machine
  * (~) extended the odom position recording to also record yaw
  * (~) rename dotproduct to dotProduct
  * (~) comment typo fixups

Like with major minor control, I've implemented some simple PID state
machines to make my life writing the auton path easier.
(Linear code calling small state machines)

The only notable thing to comment about, is how the pid move function
takes in a coordinate instead of a distance value, and how it doesn't
take in a reverse-driving predicate.

The reason for this, is for simplicity's sake, as I already have code
that records the current positions and yaw of the robot when the A
button is hit (odom.controllerUpdate).
So, I'd much like to just use the values generated from that, than
having to manually calculate it while remember orientation and
everything.

Also, just to mention, to make writing auton easier, I've also
included the current yaw in the recording of odom positions
(odom.controllerUpdate).

And, due to that, that also negates the need for a reverse flag, as,
since I use the dotproduct to find the error, (as otherwise the error
could never be reduced if the robot's yaw is off) if the goal coordinate
is behind the robot (angle difference is larger than 90*) then the robot
will automatically have a negative error, and therefore drive backwards.

Also, zig naming convention correction for the dotProduct function.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>