updated robot schematics, configs, tower & bugfixes
Changes made:
  * (~) added motor disambiguation comment for the drivetrain
  * (~) updated the ports for the drivetrain
  * (+) added proper tower motors & tower motor disambiguation comment
  * (#) fixed benchmarking to use u32 for time (ms) instead of f64
  * (+) added proper tower score high, middle, and generic spin functions
  * (+) proper L1 L2 controls for the tower
  * (+) temperature logging for all 4 of the tower motors
  * (~) updated odom imu & rotation ports
  * (~) added some explicit comptime predicates for drive control

Recently, I've been briefed with the new robot schematics, eg the
layout of motors, and sensors, and their ports, also how the motors
should be spun for the tower.

Due to the new knowledge of the layout of the drivetrain, I've written
some disambugation comments so that I know which motor is which.

Also, since the time in the benchmarking logs are going to be measured
and logged in ms, there's no need for the double-precision float, as
that's only required for storing the time in seconds
(when converted from milliseconds).

This is the functionality of the tower, as I've been briefed:
    === SCORE HIGH GOAL ===
    a: forwards
    b: backwards
    c: backwards
    d: backwards

    === SCORE MIDDLE GOAL ===
    a: off
    b: forwards
    c: backwards
    d: off
So I've written the functions to do so accordingly, with the controls
that I've decided upon being that L1, should score the high-goal,
L2 should discard the block by running a reversed version of high-goal
scoring and, if L1 & L2 are pressed at the same time, then the tower
shall score in the middle goal.

Some other port configurations were also now updated, as they are now
known (as the robot has been built).
I also noticed that the drive controls weren't explicitly making the
option predicates comptime, so there's a chance they weren't
conditionally-compiled.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>