tower: velocity -> voltage, logging velocity now
Changes made:
  * (~) switched from setting velocity to voltage for the tower motors
  * (+) added logging of the tower motor velocities in rpm
  * (~) made odom A button yaw recording display degrees

The main idea I had was, that, since the tower motors spin slower when
there's a block inside of them, if you were to graph their velocities
under a constant voltage, then you would be able to see when there was
a block in the tower, and at which motor/progress it was at.

So, if the difference is large enough, I could program auton so that
it waits for the blocks to enter the tower and exit, instead of just
using arbitrary timings.

Also, the odom coord&yaw recording, displayed radians when it should've
displayed degrees (forgot a conversion).

Signed-off-by: Ethan ma <kalscium@protonmail.com>