odom: added lateral rotation sensor with offset
Changes made:
  * (-) removed old "no lazy" test for non-lazy loading of major_minor control functions
  * (~) changed odom getRotation & getYaw functions to return null on error instead of 0
  * (+) added lateral rotation sensor & offset configs
  * (+) added functions for undoing a lateral offset (in the vertical axis)
  * (+) added lateral movement to odom updates
  * (~) separated velocity logs into lateral & vertical movement

First, I got rid of some old 'test' code for major minor control so that
the functions were actually evaluated and linted that I'd forgotten to
remove.

I've also changed the getRotation and getYaw functions to, upon error,
return null instead of 0, as there are certain scenarios in which the
function calling it, needs to know if the sensor is working or not
instead of just getting fed some placeholder value that's valid most of
the time (lateral offset correction).
The only real difference for functions that expect zero upon sensor
disconnect is the inclusion of `orelse 0`.

As the robot also has a lateral odom sensor that's offset from the
centre in the vertical (local y) axis while the lateral odom sensor
measures changes in the lateral (local x) axis, I've had to write
code to correct for this offset.
The main idea behind this offset correction, is that all movements
even lateral movements without turning move in an arc around an ICR,
and all the offset is, is a reduction in that ICR.
The ICR itself doesn't need to be calculated, only the effect of
removing the offset from it, upon the arc.

Any lateral movement is added to the coordinate vector as being a vector
shifted 90*, as that's perpendicular to it's 'local y', and it's 90 and
not -90 as an increase in the lateral movement should be an increase in
the 'local x'.

Also, the velocity of the robot is logged as vertical and lateral
velocities separately as the overall velocity can be calculated
externally and not on the robot brain (where compute is INCREDIBLY
valuable).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>