removed tower motor A
Changes made:
  * (-) tower motor A spin code
  * (-) tower motor A temp logging
  * (-) tower motor A velocity logging
  * (#) replaced %d with %lu

I've been recently notified to the removal of the top motor, motor A
from the tower.

I've just removed A, leaving the order B C D, without shifting them up.
Why? mainly because I'm lazy, and because shifting them up is prone
to human error and will be far more of a headache than it's worth.

Also, I've replaced the %d format specifier with %lu for time, as %d is
the format specifier for an i32 (twos compliement), so if the time (u32)
exceeds the i32 max value, then it would be displayed as the largest
negative number which is not what I want.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>