major minor tuning and a bunch of bugfixes
Changes made:
  * (+) added tuning of the minor movement correction
  * (+) added tuning of the minor rotation correction
  * (#) bugfixes to major_minor.zig & pid.zig from odom changes
  * (#) removed reverse from major rotations

I've added tuning for both minor controllers (mov & yaw).
This tuning code is incredibly short, as it uses the linear function
calling state-machine approach, and has no actual logging, as all of it
is just done by eye and "eh yeah, that feels about right".

Also, in creating these tuning scenarios, it evaluated some functions
that were previously lazily evaluated and therefore the errors/changes
introduced by the changes to odom (introduction of lateral rotation
sensor).
For both the PID & Major Minor state machines, I've chosen to just use
the vertical velocity value, as both these functions both can't and
shouldn't deal with any lateral velocity changes.

I was about to say that I added a 2 second delay to the tuning runtime
to allow for the IMU to calibrate, however, this should be done instead
by just using the competition switch, keeping it in disabled, waiting 2
seconds and THEN switching to driver control (for the tuning runtime).

Also, I've removed the 'reverse' flag from the major rotation controller
(major_minor.zig rotate).
This is as, the 'reverse' flag makes literally no sense in the context
of a rotation controller, and is just an artifact of the copy-paste from
the major movement controller.
The only REAL use of the reverse flag, was to check if the robot had
reached the coasting distance or overshot it.
I've replaced that functionality by simply using the sign of the
coasting displacement, as the coasting displacement takes the
sign/direction of the velocity, and the coasting distance should always
be in the same direction as the velocity (otherwise you've overshot it).

Signed-off-by: Ethan Ma <kalscium@prootonmail.com>